% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{as_raster}
\alias{as_raster}
\title{Convert MODFLOW arrays to spatial raster data}
\usage{
as_raster(array, grid, xout, yout, delta = NULL, target = NULL, crs = NA)
}
\arguments{
\item{array}{a 1D (DISV) or 2D (DIS) array}

\item{grid}{mf6grid object}

\item{xout}{numeric vector with equally spaced marginal x-coordinates of output grid}

\item{yout}{numeric vector with equally spaced marginal y-coordinates of output grid}

\item{delta}{numeric grid spacing for rectangular grid. Overwrites \code{xout} and \code{yout}.}

\item{target}{\code{terra SpatRaster} object to obtain the required dimensions from. Defaults to \code{NULL}. Overwrites \code{delta} and \code{xout} and \code{yout}}

\item{crs}{optional crs information with epsg code. Defaults to NA.}
}
\value{
An object of class \code{SpatRaster} representing the MODFLOW array
}
\description{
Convert MODFLOW arrays to spatial raster data
}
