% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{as_spatial}
\alias{as_spatial}
\title{Convert MODFLOW arrays to spatial vector data}
\usage{
as_spatial(array, grid, crs = NA, as_points = FALSE)
}
\arguments{
\item{array}{a 1D (DISV) or 2D (DIS) array}

\item{grid}{mf6grid object}

\item{crs}{optional crs information as passed to \code{\link[sf:st_crs]{sf::st_set_crs()}}. Defaults to NA.}

\item{as_points}{logical, should cell centers be returned (POINTS) or cell polygons (POLYGON, default)?}
}
\value{
A sf object of the array in POINT (\code{as_points = TRUE}) or POLYGON (\code{as_points = FALSE}) geometry with
additional columns id (cell id) and value.
}
\description{
Convert MODFLOW arrays to spatial vector data
}
