% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{as_tibble}
\alias{as_tibble}
\title{Convert a MODFLOW 6 array to a tibble}
\usage{
as_tibble(array, grid, ilay, kstp, as_points = FALSE)
}
\arguments{
\item{array}{an array with dimensions 1D, 2D or 3D (DISV) or 2D, 3D or 4D (DIS)}

\item{grid}{mf6grid object}

\item{ilay}{layer integer specifying which layer to subset}

\item{kstp}{time step integer specifying which time step to subset}

\item{as_points}{logical, should a tibble with cell center coordinates (TRUE) or cell vertices (FALSE; default) be returned?}
}
\value{
A tibble with columns id (cell id), x and y (coordinates) and value
}
\description{
\code{\link[=as_tibble]{as_tibble()}} converts a MODFLOW 6 array to a tibble, either from the cell centers or the cell polygons
}
