% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{as_tibble_cross}
\alias{as_tibble_cross}
\title{Convert a MODFLOW 6 array to a tibble along a cross-section}
\usage{
as_tibble_cross(
  array,
  grid,
  i,
  j,
  xy = NULL,
  kstp,
  as_points = FALSE,
  limit_wt = FALSE
)
}
\arguments{
\item{array}{an array with dimensions 2D or 3D (DISV) or 3D or 4D (DIS)}

\item{grid}{mf6grid object}

\item{i}{integer row index for a DIS grid to get the cross-section}

\item{j}{integer column index for a DIS grid to get the cross-section}

\item{xy}{matrix with two columns containing the x and y coordinates a cross-section line. Overwrites i and j for a DIS grid.}

\item{kstp}{time step integer specifying which time step to subset}

\item{as_points}{logical, should a tibble with cell center coordinates (TRUE)
or cell vertices (FALSE; default) be returned?}

\item{limit_wt}{logical, should the top of the cells be limited by the water table as computed from array? Defaults to \code{FALSE}.}
}
\value{
A tibble with columns id (cell id), x, y and z (coordinates), value and top and bottom elevation of the cell
}
\description{
\code{\link[=as_tibble_Cross]{as_tibble_Cross()}} converts a MODFLOW 6 array to a tibble along a vertical
cross-section, either from the cell centers or the cell polygons
}
