% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{at_water_table}
\alias{at_water_table}
\title{Get the values of an array at the water-table}
\usage{
at_water_table(array, grid, heads = NULL, kstp)
}
\arguments{
\item{array}{array with values}

\item{grid}{mf6grid object}

\item{heads}{optional secondary array containing the water-table elevations}

\item{kstp}{time step integer required for subsetting arrays with a time-dimensions}
}
\value{
The array without the final dimension with the values in the uppermost active cells (first non-NA value along vertical column)
}
\description{
\code{\link[=at_water_table]{at_water_table()}} obtains the first non-NA value along a vertical column in the array,
or based on another array
}
