% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{convert_id_to_id}
\alias{convert_id_to_id}
\title{Convert id to id for a DIS grid}
\usage{
convert_id_to_id(
  id,
  grid,
  from = c("modflow", "r"),
  to = c("r", "modflow"),
  ...
)
}
\arguments{
\item{id}{cell id, providing the place of the number in an input file 2d or
3d array}

\item{grid}{mf6grid object}

\item{from}{'r' or 'modflow'. The type of id to convert from. Defaults to
'modflow'}

\item{to}{'r' or 'modflow'. The type of id to convert to. Defaults to 'r'}

\item{...}{ignored}
}
\value{
integer vector like \code{id}, with the requested numbering type
}
\description{
\code{\link[=convert_id_to_id]{convert_id_to_id()}} converts a cell id of a DIS grid from R numbering
(column-major) to MODFLOW numbering (row-major), and vice versa.
}
\details{
a modflow id provides the place of the number in an input file 3d
array using row-major numbering; R uses column-major numbering
}
