% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{convert_id_to_ijk}
\alias{convert_id_to_ijk}
\title{Convert id to ijk for a DIS grid}
\usage{
convert_id_to_ijk(id, grid, type = c("r", "modflow"), ...)
}
\arguments{
\item{id}{cell id, providing the place of the number in an input file 3d array}

\item{grid}{mf6grid object}

\item{type}{'r' or 'modflow' specifying type of id. See details. Defaults to 'r'}

\item{...}{ignored}
}
\value{
a data.frame with column i, j and k specifying the row, column and layer indices
}
\description{
\code{\link[=convert_id_to_ijk]{convert_id_to_ijk()}} converts a cell id of a DIS grid to row-column-layer indices using R or MODFLOW numbering
}
\details{
a modflow id provides the place of the number in an input file 3d array using row-major numbering; R uses column-major numbering
}
