% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{convert_ijk_to_id}
\alias{convert_ijk_to_id}
\title{Convert ijk to id for a DIS grid}
\usage{
convert_ijk_to_id(i, j, k, grid, type = c("r", "modflow"), ...)
}
\arguments{
\item{i}{vector of row numbers}

\item{j}{vector of column numbers}

\item{k}{vector of layer numbers}

\item{grid}{mf6grid object}

\item{type}{'r' or 'modflow' specifying type of id. See details. Defaults to 'r'}

\item{...}{ignored}
}
\value{
cell ids, providing the place of the cell in an input file 3d array
}
\description{
\code{\link[=convert_ijk_to_id]{convert_ijk_to_id()}} convert row-column-layer indices for a DIS grid to cell id's using R or MODFLOW numbering
}
\details{
a modflow id provides the place of the number in an input file 3d array using row-major numbering; R uses column-major numbering
}
