% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{convert_xyz_to_grid}
\alias{convert_xyz_to_grid}
\title{Convert real-world coordinates to model cell id's, local coordinates or offsets}
\usage{
convert_xyz_to_grid(
  x,
  y,
  z = NULL,
  grid,
  output = c("id", "xyz", "offset", "all")
)
}
\arguments{
\item{x}{numeric vector with real-world x-coordinate}

\item{y}{numeric vector with real-world y-coordinate}

\item{z}{optional numeric vector with real-world z-coordinate}

\item{grid}{mf6grid object}

\item{output}{character defining the output type, either \code{"id"} (for cell id's), \code{"xyz"} for local coordinates, \code{"offset"} for cell offsets or \code{"all"} for all of the above.}
}
\value{
For \code{output = "id"} a vector with the cell id for the cell containing the corresponding point. Points that fall on a cell boundary will return the largest cell id value.
Points outside the model grid will return \code{NA}.

It is assumed the supplied real-world coordinates are in the same coordinate system as \code{grid}.
}
\description{
\code{\link[=convert_xyz_to_grid]{convert_xyz_to_grid()}} converts real-world x-y (and optionally z) coordinates to cell id's,
local grid coordinates and/or offsets for a MODFLOW 6 grid
}
