% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/etl.R
\name{csv2db}
\alias{csv2db}
\title{An alternative database loading mechanism that relies on having a
directory full of csv files.}
\usage{
csv2db(db, path)
}
\arguments{
\item{db}{a database connection as initialized by src_dbi from
dplyr}

\item{path}{a path to a bunch of csv files that should be loaded
into a db}
}
\value{
list of table handles, side effect is a database written to disk
}
\description{
An alternative database loading mechanism that relies on having a
directory full of csv files.
}
\examples{
foo.db <- src_postgres("foo")
csv2db(foo.db, "~/path/to/csvs")
}

