% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkbuild.R
\name{edge.attrs}
\alias{edge.attrs}
\title{Identifies edges between table attributes based on arbitrary
function}
\usage{
edge.attrs(attrs, fn = intersect, tabs = FALSE, ...)
}
\arguments{
\item{attrs}{a list containing table attributes, as returned by
ls.attr}

\item{fn}{function used to compare attribute names, returning
logical intersection}

\item{tabs}{does the fn require table names to be passed
(e.g. function works against a sql store)?}
}
\value{
dataframe representing an edgelist
}
\description{
Identifies edges between table attributes based on arbitrary
function
}
\examples{
foo.db <- src_postgres("foo")
attrs <- ls.attrs(foo.db)

res.grep <- edge.attrs(attrs, fn=intersect.grep)
res.token <- edge.attrs(attrs, fn=intersect.token)
res.innerjoin <- edge.attrs(attrs, fn=intersect.innerjoin, tabs=TRUE, db=foo.db)

library(igraph)
graph.grep <- graph.data.frame(res.grep, directed=FALSE)
write.graph(graph.grep, "~/Desktop/graph.grep.dot", format="dot")
plot(graph) # might take a while!

res.grep.n <- res.grep \%>\%
                 group_by(xn, yn) \%>\%
                 summarize(count=n())
graph.grep.n <- graph.data.frame(res.grep.n, directed=FALSE)
}

