% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkbuild.R
\name{intersect.agrep}
\alias{intersect.agrep}
\title{Intersect based on agrep. Much slower / more inefficient than
base::intersect. Typically called by edge.attrs()}
\usage{
intersect.agrep(xs, ys, ...)
}
\value{
a vector containing the intersection of xs and ys where
matching elements are sorted and concatenated to form an identifier
}
\description{
Intersect based on agrep. Much slower / more inefficient than
base::intersect. Typically called by edge.attrs()
}

