% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkbuild.R
\name{intersect.innerjoin}
\alias{intersect.innerjoin}
\title{Intersect based on innerjoin of table crossproducts (VERY SLOW)}
\usage{
intersect.innerjoin(xs, ys, xtab, ytab, db, threshold = 1, ...)
}
\arguments{
\item{xs}{a vector of strings to compare}

\item{ys}{a vector of strings to compare}

\item{xtab}{name of the table containing xs}

\item{ytab}{name of the table containing ys}

\item{db}{database connection as set up by dplyr}

\item{threshold}{optional, number of matching rows to constitute a
column match}
}
\description{
Intersect based on innerjoin of table crossproducts (VERY SLOW)
}

