% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkbuild.R
\name{intersect.token}
\alias{intersect.token}
\title{Intersect based on tokenized versions of names}
\usage{
intersect.token(xs, ys, ...)
}
\arguments{
\item{xs}{a vector of strings to compare}

\item{ys}{a vector of strings to compare}
}
\value{
a vector containing the intersection of xs and ys where
matching elements are sorted and concatenated to form an identifier
}
\description{
Intersect based on tokenized versions of names
}

