% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkbuild.R
\name{ls.attr}
\alias{ls.attr}
\title{Function lists the attributes of one or more tables}
\usage{
ls.attr(db)
}
\arguments{
\item{db}{connection to a database as set up by dplyr}

\item{tbids}{character vector of table identifiers}
}
\value{
data.frame
}
\description{
Function lists the attributes of one or more tables
}
\examples{
foo.db <- src_postgres("foo")
attrs <- ls.attr(foo.db)
}

