% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/etl.R
\name{mk.socrata.url}
\alias{mk.socrata.url}
\title{Construct a socrata url from a site root and an identifier}
\usage{
mk.socrata.url(site, identifier)
}
\arguments{
\item{site}{root of socrata site (e.g. "https://data.seattle.gov")}

\item{identifier}{}
}
\description{
Construct a socrata url from a site root and an identifier
}

