% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/modr.R
\docType{package}
\name{modr}
\alias{modr}
\alias{modr-package}
\title{modr: a package for reproducing the municipal open data project's
analysis.}
\description{
Package provides:
}
\details{
- extract, transform, and load (ETL) procedures for getting data
  out of socrata and into a local DB.

- procedures for identifying linkages between datasets

- procedures for visualizing analysis
}

