% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{read_concentrations}
\alias{read_concentrations}
\title{Read a MODFLOW 6 output concentration file}
\usage{
read_concentrations(file, grid, totim = NULL, set_na = TRUE)
}
\arguments{
\item{file}{path to the output file containing the simulated concentrations}

\item{grid}{mf6grid object}

\item{totim}{optional numeric vector with simulation times at which output is to be read}

\item{set_na}{logical, should dry and inactive cells be set to NA? Defaults to TRUE}
}
\value{
Array with the concentration values. The dimensions are NROW NCOL NLAY NSTPS for DIS,
NCPL NLAY NSTPS for DISV and NCELLS NSTPS for DISU
}
\description{
\code{\link[=read_concentration]{read_concentration()}} reads a MODFLOW 6 output concentration file
}
