% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{read_grid}
\alias{read_grid}
\title{Read a MODFLOW 6 binary grid file}
\usage{
read_grid(file)
}
\arguments{
\item{file}{path to the binary grid file}
}
\value{
An object of class mf6grid which is a list containing the grid information for the MODFLOW 6 model
}
\description{
\code{\link[=read_grid]{read_grid()}} reads a MODFLOW 6 binary grid file and returns it as a mf6grid object
}
