% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{read_heads}
\alias{read_heads}
\title{Read a MODFLOW 6 output head file}
\usage{
read_heads(file, grid, totim = NULL, set_na = TRUE)
}
\arguments{
\item{file}{path to the output file containing the simulated heads}

\item{grid}{mf6grid object}

\item{totim}{optional numeric vector with simulation times at which output is to be read}

\item{set_na}{logical, should dry and inactive cells be set to NA? Defaults to TRUE}
}
\value{
Array with the head values. The dimensions are NROW NCOL NLAY NSTPS for DIS,
NCPL NLAY NSTPS for DISV and NCELLS NSTPS for DISU
}
\description{
\code{\link[=read_head]{read_head()}} reads a MODFLOW 6 output head file
}
