% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmf6.R
\name{resample_to_grid}
\alias{resample_to_grid}
\title{Resample values on a MODFLOW grid to a rectilinear output grid}
\usage{
resample_to_grid(array, grid, xout, yout)
}
\arguments{
\item{array}{a 1D (DISV) or 2D (DIS) array}

\item{grid}{mf6grid object}

\item{xout}{numeric vector with marginal x-coordinates of output grid}

\item{yout}{numeric vector with marginal y-coordinates of output grid}
}
\value{
Matrix with interpolated values onto the output grid.
}
\description{
Resample values on a MODFLOW grid to a rectilinear output grid
}
