% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/etl.R
\name{socrata2db}
\alias{socrata2db}
\title{Import data from remote socrata server to local sqlite
database. Note, this isn't the most memory efficient approach to
populating a db. Big datasets may choke this. But, on the other
hand it should be able to infer column datatypes.}
\usage{
socrata2db(db, site = "https://data.seattle.gov", app_token = NULL)
}
\arguments{
\item{db}{a database connection as initialized by src_dbi from
dplyr}

\item{site}{root of socrata site (e.g. "https://data.seattle.gov")}
}
\value{
list of table handles, side effect is a database written to
disk
}
\description{
Import data from remote socrata server to local sqlite
database. Note, this isn't the most memory efficient approach to
populating a db. Big datasets may choke this. But, on the other
hand it should be able to infer column datatypes.
}
\examples{
foo.db <- src_postgres("foo")
socrata2db(foo.db)
}

