% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/networkbuild.R
\name{tokify}
\alias{tokify}
\title{Converts a string to tokens}
\usage{
tokify(x, split = NULL)
}
\arguments{
\item{x}{a string}

\item{split}{optional, regex to split on}
}
\description{
Converts a string to tokens
}

